refactor: moved components to components folder
