+++ src/Components/Experience/ExperienceView.js
@@ -5,7 +5,7 @@ import Row from "react-bootstrap/Row";
 import Details from "./Components/Details";
 
 import { experienceData } from "../Data/Experience";
-import "../style/custom.css";
+import "../../style/custom.css";
 
 import Fade from "react-reveal/Fade";
