+++ src/Components/ExtraActivites/components/ActivitySlider.js
@@ -1,13 +1,13 @@
-import React, { Component } from 'react';
-import Slider from 'react-slick';
+import React, { Component } from "react";
+import Slider from "react-slick";
 
-import Col from 'react-bootstrap/Col';
+import Col from "react-bootstrap/Col";
 
-import ironmanPhoto from '../../Images/ironman.jpg';
-import wrestlingClub from '../../Images/wrestling-club.jpg';
-import samaritansPurse from '../../Images/samaritans-purse-group.jpg';
-import mountains from '../../Images/mountaineering.jpg';
-import highSchoolWrestling from '../../Images/high-school-wrestling.jpg';
+import ironmanPhoto from "../../../Images/ironman.jpg";
+import wrestlingClub from "../../../Images/wrestling-club.jpg";
+import samaritansPurse from "../../../Images/samaritans-purse-group.jpg";
+import mountains from "../../../Images/mountaineering.jpg";
+import highSchoolWrestling from "../../../Images/high-school-wrestling.jpg";
 
 export default class SimpleSlider extends Component {
   render() {
@@ -21,7 +21,7 @@ export default class SimpleSlider extends Component {
 
     const style = {
       image: {
-        width: '100%'
+        width: "100%"
       }
     };
