+++ src/Components/Projects/components/Project.js
@@ -3,7 +3,7 @@ import React from "react";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 
-import "../../style/custom.css";
+import "../../../style/custom.css";
 
 const Project = props => {
   const { projectData, indexNumber } = props;
